requirements.txt
instaloader>=4.10
pytesseract>=0.3.10
Pillow>=10.0.0
python-dateutil>=2.8.2
regex>=2023.10.3
requests>=2.31.0
